pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["RKV102 <roman.krilow@mail.ru>"]
readme = "README.md"
packages = [
  { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
django-extensions = "^3.2.3"
dj-database-url = "^2.2.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-bootstrap5 = "^24.2"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
